+++ src/Model/File.php
@@ -3,7 +3,7 @@
 namespace Couscous\Model;
 
 /**
- * Represents a file of the website.
+ * Represents a file of the repository.
  *
  * @author Matthieu Napoli <matthieu@mnapoli.fr>
  */
@@ -16,5 +16,19 @@ abstract class File
         $this->relativeFilename = $relativeFilename;
     }
 
+    /**
+     * Returns the content of the file.
+     *
+     * @return string
+     */
     public abstract function getContent();
+
+    /**
+     * Returns an indexed array of metadata.
+     *
+     * Array keys are metadata names, array values are metadata values.
+     *
+     * @return Metadata
+     */
+    public abstract function getMetadata();
 }