+++ src/Model/File/MarkdownFile.php
@@ -1,6 +1,9 @@
 <?php
 
-namespace Couscous\Model;
+namespace Couscous\Model\File;
+
+use Couscous\Model\File;
+use Couscous\Model\Metadata;
 
 /**
  * @author Matthieu Napoli <matthieu@mnapoli.fr>
@@ -12,15 +15,26 @@ class MarkdownFile extends File
      */
     public $content;
 
+    /**
+     * @var Metadata
+     */
+    private $metadata;
+
     public function __construct($relativeFilename, $content)
     {
         parent::__construct($relativeFilename);
 
         $this->content = $content;
+        $this->metadata = new Metadata();
     }
 
     public function getContent()
     {
         return $this->content;
+    }
+
+    public function getMetadata()
+    {
+        return $this->metadata;
     }
 }