+++ src/Step/Template/ProcessTwigLayouts.php
@@ -2,7 +2,7 @@
 
 namespace Couscous\Step\Template;
 
-use Couscous\Model\HtmlFile;
+use Couscous\Model\File\HtmlFile;
 use Couscous\Model\Repository;
 use Couscous\Step\StepInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -22,25 +22,25 @@ class ProcessTwigLayouts implements StepInterface
 
     public function __invoke(Repository $repository, OutputInterface $output)
     {
-        if (! $repository->template) {
+        if (! $repository->metadata['template.directory']) {
             return;
         }
 
-        $twig = $this->createTwig($repository->template->directory);
+        $twig = $this->createTwig($repository->metadata['template.directory']);
 
         /** @var HtmlFile[] $htmlFiles */
-        $htmlFiles = $repository->findFilesByType('Couscous\Model\HtmlFile');
+        $htmlFiles = $repository->findFilesByType('Couscous\Model\File\HtmlFile');
 
         foreach ($htmlFiles as $file) {
-            $layout = isset($file->customVariables['layout'])
-                ? $file->customVariables['layout'] . '.twig'
+            $fileMetadata = $file->getMetadata();
+            $layout = isset($fileMetadata['layout'])
+                ? $fileMetadata['layout'] . '.twig'
                 : self::DEFAULT_LAYOUT_NAME;
 
             $context = array_merge(
-                $repository->template->layoutVariables,
-                $repository->config->templateVariables,
-                $file->customVariables,
-                array('content' => $file->content)
+                $repository->metadata->toArray(),
+                $fileMetadata->toArray(),
+                ['content' => $file->content]
             );
 
             try {